Reshape nidx_paragraph to nidx structure
