+++ nidx/nidx_paragraph/Cargo.toml
@@ -9,6 +9,7 @@ itertools = "0.13.0"
 lazy_static = "1.5.0"
 levenshtein_automata = "0.2.1"
 nidx_protos = { version = "0.1.0", path = "../nidx_protos" }
+nidx_tantivy = { version = "0.1.0", path = "../nidx_tantivy" }
 nidx_types = { version = "0.1.0", path = "../nidx_types" }
 once_cell = "1.20.2"
 regex = "1.11.1"
