+++ nidx/nidx_text/src/lib.rs
@@ -21,13 +21,13 @@
 pub mod prefilter;
 mod query_io;
 mod reader;
+mod resource_indexer;
 mod schema;
 mod search_query;
 
 use std::collections::HashSet;
 use std::path::Path;
 
-use anyhow::anyhow;
 use nidx_protos::resource::ResourceStatus;
 use nidx_protos::{DocumentSearchRequest, DocumentSearchResponse, StreamRequest};
 use nidx_tantivy::index_reader::{open_index_with_deletions, DeletionQueryBuilder};
@@ -35,16 +35,15 @@ use nidx_tantivy::{TantivyIndexer, TantivyMeta, TantivySegmentMetadata};
 use nidx_types::OpenIndexMetadata;
 use prefilter::{PreFilterRequest, PreFilterResponse};
 use reader::{DocumentIterator, TextReaderService};
-use schema::{timestamp_to_datetime_utc, TextSchema};
+use resource_indexer::index_document;
+use schema::TextSchema;
 
 pub use search_query::TextContext;
 use tantivy::indexer::merge_indices;
 use tantivy::schema::Field;
 use tantivy::{
     directory::MmapDirectory,
-    doc,
     query::{Query, TermSetQuery},
-    schema::Facet,
     Term,
 };
 
@@ -99,70 +98,6 @@ impl TextIndexer {
             },
         })
     }
-}
-
-fn index_document(
-    writer: &mut TantivyIndexer,
-    resource: &nidx_protos::Resource,
-    schema: TextSchema,
-) -> anyhow::Result<usize> {
-    let Some(resource_id) = resource.resource.as_ref().map(|r| r.uuid.as_str()) else {
-        return Err(anyhow!("Missing resource ID"));
-    };
-    let Some(metadata) = resource.metadata.as_ref() else {
-        return Err(anyhow!("Missing resource metadata"));
-    };
-    let Some(modified) = metadata.modified.as_ref() else {
-        return Err(anyhow!("Missing resource modified date in metadata"));
-    };
-    let Some(created) = metadata.created.as_ref() else {
-        return Err(anyhow!("Missing resource created date in metadata"));
-    };
-
-    let mut base_doc = doc!(
-        schema.uuid => resource_id.as_bytes(),
-        schema.modified => timestamp_to_datetime_utc(modified),
-        schema.created => timestamp_to_datetime_utc(created),
-        schema.status => resource.status as u64,
-    );
-
-    let resource_security = resource.security.as_ref();
-    if let Some(security_groups) = resource_security.filter(|i| !i.access_groups.is_empty()) {
-        base_doc.add_u64(schema.groups_public, 0_u64);
-        for group_id in security_groups.access_groups.iter() {
-            let mut group_id_key = group_id.clone();
-            if !group_id.starts_with('/') {
-                // Slash needs to be added to be compatible with tantivy facet fields
-                group_id_key = "/".to_string() + group_id;
-            }
-            let facet = Facet::from(group_id_key.as_str());
-            base_doc.add_facet(schema.groups_with_access, facet)
-        }
-    } else {
-        base_doc.add_u64(schema.groups_public, 1_u64);
-    }
-
-    for label in resource.labels.iter() {
-        let facet = Facet::from(label.as_str());
-        base_doc.add_facet(schema.facets, facet);
-    }
-
-    for (field, text_info) in &resource.texts {
-        let mut field_doc = base_doc.clone();
-        let mut facet_key: String = "/".to_owned();
-        facet_key.push_str(field.as_str());
-        let facet_field = Facet::from(facet_key.as_str());
-        field_doc.add_facet(schema.field, facet_field);
-        field_doc.add_text(schema.text, &text_info.text);
-
-        for label in text_info.labels.iter() {
-            let facet = Facet::from(label.as_str());
-            field_doc.add_facet(schema.facets, facet);
-        }
-        writer.add_document(field_doc).unwrap();
-    }
-
-    Ok(resource.texts.len())
 }
 
 pub struct TextSearcher {
